add explicite cast from double to size_t to get rid of warning message

+++ SeedMaker.cpp
@@ -65,8 +65,8 @@ void
 SeedMaker::first_last_seeds(const size_t read_width, const size_t n_seeds, 
 			    const size_t seed_weight, vector<size_t> &profs) {
   
-  const size_t shift = 
-    std::ceil(static_cast<double>(read_width - seed_weight)/(n_seeds - 1));
+  const size_t shift = static_cast<size_t>(
+      std::ceil(static_cast<double>(read_width - seed_weight)/(n_seeds - 1)));
   const SeedMaker sm_first(read_width, n_seeds,  seed_weight, seed_weight,
 			   shift, read_width);
   const SeedMaker sm_last(read_width, n_seeds, seed_weight, 1, 1, n_seeds);
@@ -104,8 +104,8 @@ SeedMaker::last_seeds(const size_t read_width, const size_t n_seeds,
 void
 SeedMaker::first_seeds(const size_t read_width, const size_t n_seeds, 
 		       const size_t seed_weight, vector<size_t> &profs) {
-  const size_t shift = 
-    std::ceil(static_cast<double>(read_width - seed_weight)/(n_seeds - 1));
+  const size_t shift = static_cast<size_t>(
+      std::ceil(static_cast<double>(read_width - seed_weight)/(n_seeds - 1)));
   const SeedMaker sm_first(read_width, n_seeds,  seed_weight, seed_weight,
 			   shift, read_width);
   sm_first.get_seed_profiles(profs);
